Replace `[ExpectedException]` with `Assert.Throws`
As outlined in issue #34, unit tests relying on a `[ExpectedException]`
attribute should be rewritten to use `Assert.Throws` instead:

 * `[ExpectedException]` is too coarse-grained and doesn't allow us to
   test only the Act stage in Arrange-Act-Assert. `Assert.Throws` does
   not suffer from this.

 * `[ExpectedException]` is no longer supported in NUnit version 3.

This commit replaces all occurrences of the attribute with the more
flexible `Assert.Throws`. In some cases, parts of the `TestDelegate`
that belong to the Arrange stage are moved out of it.