Add option to throw on calling RMS.ToArray()
Implementations can reference an RMS object as MemoryStream - which may internally call ToArray(); this is common behavior for expandable MemoryStreams. As of now, this can only be found by looking at the ETW log. An easier way is to throw an exception on a call to ToArray, which is set by the  RecyclableMemoryStreamManager.ThrowOnCallToArray option.